refactor: PR 리뷰 반영    - 서비스에서 HttpException 정적 메서드 사용    - 파일명 kebab-case 컨벤션 통일    - import 경로 업데이트
